+++ packages/devextreme/js/__internal/ui/color_box/color_box.ts
@@ -8,8 +8,8 @@ import Color from '@ts/m_color';
 import DropDownEditor from '@ts/ui/drop_down_editor/drop_down_editor';
 import type { ValueChangedEvent } from '@ts/ui/editor/editor';
 
-import type { PopupProperties } from '../popup/m_popup';
-import type Popup from '../popup/m_popup';
+import type { PopupProperties } from '../popup/popup';
+import type Popup from '../popup/popup';
 import type { ColorViewProperties } from './color_view';
 import ColorView from './color_view';
