+++ packages/devextreme/js/__internal/ui/date_box/m_date_box.strategy.calendar_with_time.ts
@@ -7,7 +7,7 @@ import { getWindow } from '@js/core/utils/window';
 import Box from '@ts/ui/box';
 import TimeView from '@ts/ui/date_box/time_view';
 
-import type { PopupProperties } from '../popup/m_popup';
+import type { PopupProperties } from '../popup/popup';
 import type DateBox from './date_box.base';
 import uiDateUtils from './date_utils';
 import CalendarStrategy from './m_date_box.strategy.calendar';
