+++ packages/devextreme/js/__internal/ui/diagram/ui.diagram.panel.ts
@@ -5,7 +5,8 @@ import type { dxElementWrapper } from '@js/core/renderer';
 import $ from '@js/core/renderer';
 import type { OptionChanged } from '@ts/core/widget/types';
 import Widget from '@ts/core/widget/widget';
-import type { PopupProperties } from '@ts/ui/popup/m_popup';
+
+import type { PopupProperties } from '../popup/popup';
 
 const POINTERUP_EVENT_NAME = addNamespace(pointerEvents.up, 'dxDiagramPanel');
 const PREVENT_REFOCUS_SELECTOR = '.dx-textbox';
