+++ packages/devextreme/js/__internal/ui/diagram/ui.diagram.toolbox.ts
@@ -11,9 +11,10 @@ import type { dxPopupAnimation } from '@js/ui/popup';
 import { getDiagram } from '@ts/ui/diagram/diagram.importer';
 import DiagramFloatingPanel from '@ts/ui/diagram/ui.diagram.floating_panel';
 import Tooltip from '@ts/ui/m_tooltip';
-import type { PopupProperties } from '@ts/ui/popup/m_popup';
 import ScrollView from '@ts/ui/scroll_view/scroll_view';
 import TextBox from '@ts/ui/text_box/text_box';
+
+import type { PopupProperties } from '../popup/popup';
 
 const DIAGRAM_TOOLBOX_MIN_HEIGHT = 130;
 const DIAGRAM_TOOLBOX_POPUP_CLASS = 'dx-diagram-toolbox-popup';
