+++ packages/devextreme/js/__internal/ui/dialog.ts
@@ -22,8 +22,9 @@ import type { ToolbarItem } from '@js/ui/popup';
 import { current, isFluent } from '@js/ui/themes';
 import errors from '@js/ui/widget/ui.errors';
 import domUtils from '@ts/core/utils/m_dom';
-import type { PopupProperties } from '@ts/ui/popup/m_popup';
-import Popup from '@ts/ui/popup/m_popup';
+
+import type { PopupProperties } from './popup/popup';
+import Popup from './popup/popup';
 
 interface BaseDialog {
   show: () => Promise<unknown>;
