+++ packages/devextreme/js/__internal/ui/popup/m_popup.full.ts
@@ -2,7 +2,8 @@ import '@js/ui/toolbar';
 
 import registerComponent from '@js/core/component_registrator';
 import Popup from '@js/ui/popup/ui.popup';
-import type { PopupProperties } from '@ts/ui/popup/m_popup';
+
+import type { PopupProperties } from './popup';
 
 const TOOLBAR_WIDGET_NAME = 'dxToolbar';
