+++ packages/devextreme/js/__internal/ui/popup/popup.ts
@@ -320,7 +320,6 @@ class Popup<
     ];
   }
 
-  // eslint-disable-next-line class-methods-use-this
   _iosAnimation(): dxPopupAnimation {
     return {
       show: {
@@ -721,7 +720,6 @@ class Popup<
     return $toolbar;
   }
 
-  // eslint-disable-next-line class-methods-use-this
   _renderByTemplate(
     template: TemplateBase,
     $container: dxElementWrapper,
@@ -841,7 +839,7 @@ class Popup<
       // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
       item.toolbar = data.toolbar || item.toolbar || 'top';
 
-      if (item && item.toolbar === toolbar) {
+      if (item?.toolbar === toolbar) {
         if (isShortcut) {
           extend(item, { location: data.location }, this._getToolbarItemByAlias(data));
         }
@@ -934,7 +932,6 @@ class Popup<
     return action;
   }
 
-  // eslint-disable-next-line class-methods-use-this
   _getToolbarName(): string {
     return TOOLBAR_BASE_WIDGET_NAME;
   }
@@ -1248,7 +1245,6 @@ class Popup<
     return cssStyles;
   }
 
-  // eslint-disable-next-line class-methods-use-this
   _setHeightClasses($container: dxElementWrapper, currentClass: string): void {
     let excessClasses = '';
