+++ packages/devextreme/js/__internal/ui/popup/popup_position_controller.ts
@@ -11,7 +11,8 @@ import type {
   PositionControllerConstructor,
 } from '@ts/ui/overlay/overlay_position_controller';
 import { OverlayPositionController } from '@ts/ui/overlay/overlay_position_controller';
-import type { PopupProperties } from '@ts/ui/popup/m_popup';
+
+import type { PopupProperties } from './popup';
 
 export interface PopupControllerProperties extends BaseControllerProperties {
   fullScreen?: PopupProperties['fullScreen'];
