+++ packages/devextreme/testing/tests/DevExpress.ui.widgets/chatParts/confirmationPopup.markup.tests.js
@@ -6,7 +6,7 @@ import ConfirmationPopup, {
 } from '__internal/ui/chat/confirmationpopup';
 
 import { OVERLAY_CONTENT_CLASS } from '__internal/ui/overlay/overlay';
-import { POPUP_CONTENT_CLASS } from '__internal/ui/popup/m_popup';
+import { POPUP_CONTENT_CLASS } from '@ts/ui/popup/popup';
 import { BUTTON_CLASS } from '__internal/ui/button/button';
 
 const moduleConfig = {
